+++ Source/Core/DoubleRange.cs
@@ -7,53 +7,53 @@ namespace SharpMSDF.Core
     /// </summary>
     public struct DoubleRange
     {
-        public double Lower, Upper;
+        public float Lower, Upper;
 
-        public DoubleRange(double symmetricalWidth = 0)
+        public DoubleRange(float symmetricalWidth = 0)
         {
-            Lower = -0.5 * symmetricalWidth;
-            Upper = 0.5 * symmetricalWidth;
+            Lower = -0.5f * symmetricalWidth;
+            Upper = 0.5f * symmetricalWidth;
         }
 
-        public DoubleRange(double lowerBound, double upperBound)
+        public DoubleRange(float lowerBound, float upperBound)
         {
             Lower = lowerBound;
             Upper = upperBound;
         }
 
-        public static DoubleRange operator *(DoubleRange range, double factor)
+        public static DoubleRange operator *(DoubleRange range, float factor)
         {
             return new DoubleRange(range.Lower * factor, range.Upper * factor);
         }
 
-        public static DoubleRange operator *(double factor, DoubleRange range)
+        public static DoubleRange operator *(float factor, DoubleRange range)
         {
             return new DoubleRange(factor * range.Lower, factor * range.Upper);
         }
 
-        public static DoubleRange operator /(DoubleRange range, double divisor)
+        public static DoubleRange operator /(DoubleRange range, float divisor)
         {
             return new DoubleRange(range.Lower / divisor, range.Upper / divisor);
         }
 
-        public void MultiplyInPlace(double factor)
+        public void MultiplyInPlace(float factor)
         {
             Lower *= factor;
             Upper *= factor;
         }
 
-        public void DivideInPlace(double divisor)
+        public void DivideInPlace(float divisor)
         {
             Lower /= divisor;
             Upper /= divisor;
         }
 
-        public DoubleRange Multiply(double factor)
+        public DoubleRange Multiply(float factor)
         {
             return new DoubleRange(Lower * factor, Upper * factor);
         }
 
-        public DoubleRange Divide(double divisor)
+        public DoubleRange Divide(float divisor)
         {
             return new DoubleRange(Lower / divisor, Upper / divisor);
         }
